dapp-refactor: moved /hooks folder to /src folder
