+++ system-contract-dapp-playground/src/components/common/components/OneLineMethod.tsx
