+++ system-contract-dapp-playground/src/components/contract-interaction/erc/erc-20/methods/mint/index.tsx
@@ -24,12 +24,12 @@ import { useEffect, useState } from 'react';
 import { Contract, isAddress } from 'ethers';
 import { erc20Mint } from '@/api/hedera/erc20-interactions';
 import { CommonErrorToast } from '@/components/toast/CommonToast';
-import MultiLineMethod from '@/components/common/MultiLineMethod';
 import { TransactionResult } from '@/types/contract-interactions/HTS';
+import MultiLineMethod from '@/components/common/components/MultiLineMethod';
 import { mintParamFields } from '@/utils/contract-interactions/erc/erc20/constant';
 import { CONTRACT_NAMES, HEDERA_TRANSACTION_RESULT_STORAGE_KEYS } from '@/utils/common/constants';
 import { handleAPIErrors } from '@/components/contract-interaction/hts/shared/methods/handleAPIErrors';
-import { useUpdateTransactionResultsToLocalStorage } from '@/components/contract-interaction/hts/shared/hooks/useUpdateLocalStorage';
+import { useUpdateTransactionResultsToLocalStorage } from '@/hooks/useUpdateLocalStorage';
 
 interface PageProps {
   baseContract: Contract;
