+++ system-contract-dapp-playground/src/components/contract-interaction/erc/erc-721/methods/owner-of/index.tsx
@@ -23,7 +23,7 @@ import { Contract } from 'ethers';
 import { ReactNode, useCallback, useState } from 'react';
 import { erc721OwnerOf } from '@/api/hedera/erc721-interactions';
 import { CommonErrorToast } from '@/components/toast/CommonToast';
-import HederaCommonTextField from '@/components/common/HederaCommonTextField';
+import HederaCommonTextField from '@/components/common/components/HederaCommonTextField';
 import { Table, TableContainer, Tbody, Th, Thead, Tr, useToast } from '@chakra-ui/react';
 import useUpdateMapStateUILocalStorage from '../../../shared/hooks/useUpdateMapStateUILocalStorage';
 import useRetrieveMapValueFromLocalStorage from '../../../shared/hooks/useRetrieveMapValueFromLocalStorage';
