+++ system-contract-dapp-playground/src/components/contract-interaction/exchange-rate/deployment/ExchangeRateDeployField.tsx
@@ -22,7 +22,7 @@ import Image from 'next/image';
 import { useToast } from '@chakra-ui/react';
 import { Dispatch, SetStateAction, useState } from 'react';
 import { CommonErrorToast } from '../../../toast/CommonToast';
-import HederaCommonTextField from '../../../common/HederaCommonTextField';
+import HederaCommonTextField from '../../../common/components/HederaCommonTextField';
 
 interface PageProps {
   isDeploying: boolean;
@@ -30,11 +30,7 @@ interface PageProps {
   setDidDeployStart: Dispatch<SetStateAction<boolean>>;
 }
 
-const ExchangeRateDeployField = ({
-  isDeploying,
-  setDidDeployStart,
-  setDeployedParams,
-}: PageProps) => {
+const ExchangeRateDeployField = ({ isDeploying, setDidDeployStart, setDeployedParams }: PageProps) => {
   const toaster = useToast();
   const [value, setValue] = useState('');
