+++ system-contract-dapp-playground/src/components/contract-interaction/index.tsx
@@ -25,16 +25,16 @@ import { BsTrash } from 'react-icons/bs';
 import HederaPRNGMethods from './prng/methods';
 import { FiExternalLink } from 'react-icons/fi';
 import ERC20Methods from './erc/erc-20/methods';
-import ConfirmModal from '../common/ConfirmModal';
 import ERC721Methods from './erc/erc-721/methods';
 import HederaIHRC719Methods from './ihrc/methods';
 import { deploySmartContract } from '@/api/hedera';
-import HederaAlertDialog from '../common/AlertDialog';
 import { useCallback, useEffect, useState } from 'react';
 import { generateBaseContractInstance } from '@/api/ethers';
+import ConfirmModal from '../common/components/ConfirmModal';
 import { clearCachedTransactions } from '@/api/localStorage';
 import ERC20DeployField from './erc/deployment/ERCDeployField';
 import HederaExchangeRateMethods from './exchange-rate/methods';
+import HederaAlertDialog from '../common/components/AlertDialog';
 import { HederaContractAsset, NetworkName } from '@/types/common';
 import { getHederaNativeIDFromEvmAddress } from '@/api/mirror-node';
 import { CommonErrorToast, NoWalletToast } from '../toast/CommonToast';
