+++ system-contract-dapp-playground/src/components/wallet-popup/index.tsx
@@ -23,8 +23,8 @@ import Image from 'next/image';
 import { ethers } from 'ethers';
 import { clearCookies } from '@/api/cookies';
 import { NetworkName } from '@/types/common';
-import ConfirmModal from '../common/ConfirmModal';
 import { BiCopy, BiCheckDouble } from 'react-icons/bi';
+import ConfirmModal from '../common/components/ConfirmModal';
 import { clearCachedTransactions } from '@/api/localStorage';
 import { getBalance, getWalletProvider } from '@/api/wallet';
 import { getHederaNativeIDFromEvmAddress } from '@/api/mirror-node';
